Updating Mobile expenses app now that prices are stored in dollars instead of cents.
Review at http://gwt-code-reviews.appspot.com/514801

Review by: jgw@google.com

git-svn-id: http://google-web-toolkit.googlecode.com/svn/branches/2.1@8108 8db76d5a-ed1c-0410-87a9-c151d255dfc7

+++ bikeshed/src/com/google/gwt/sample/expenses/gwt/client/MobileExpenseDetails.java
@@ -119,6 +119,6 @@ public class MobileExpenseDetails extends Composite implements MobilePage {
     nameText.setInnerText(expense.getDescription());
     dateText.setInnerText(formatter.format(expense.getCreated()));
     categoryText.setInnerText(expense.getCategory());
-    priceText.setInnerText(ExpensesMobile.formatCurrency(expense.getAmount().intValue()));
+    priceText.setInnerText(ExpensesMobile.formatCurrency(expense.getAmount()));
   }
 }
